requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
gevent-websocket==0.10.1
irc==20.1.0
python-dotenv==0.19.0
Werkzeug==2.2.3
openai>=1.12.0
requests
beautifulsoup4
urllib3
google-generativeai>=0.3.2
anyio>=4.8.0,<5.0.0
SQLAlchemy==1.4.46 